Support for custom lattice function
A custom lattice function can now be specified on SolverPaths to give the user control over the initial ray directions. While a fibonacci lattice is useful in e.g. urban situations where radiation should be isotropic, in scenarios such as for satellite links a custom lattice might be better suited.